+++ src/main/java/com/enhancer/planner/ProposalState.java
@@ -1,0 +1,5 @@
+package com.enhancer.planner;
+
+public enum ProposalState {
+    PROPOSAL
+}